# agent/openapi/week2.yaml
components:
  schemas:
...
    AllergyWrite:
      description: 'An allergy (issue) list entry, appended as a new list row.


        ``title`` is the intake-derived allergen/substance string the physician

        confirms. ``begdate`` is a ``YYYY-MM-DD`` string, format-checked by the

        write verification step.'
      properties:
        begdate:
          description: Onset/entry date, YYYY-MM-DD.
          minLength: 1
          title: Begdate
          type: string
        reaction:
          anyOf:
          - type: string
          - type: 'null'
          description: Optional reaction description.
          title: Reaction
        title:
          minLength: 1
          title: Title
          type: string
      required:
      - title
      - begdate
      title: AllergyWrite
      type: object
...
    ClinicianId:
      description: OpenEMR user id — positive integer only.
      properties:
        value:
          exclusiveMinimum: 0.0
          title: Value
          type: integer
      required:
      - value
      title: ClinicianId
      type: object
    ConfirmRequest:
      description: 'The confirm transaction: the identical candidate echoed back verbatim.


        The candidate carries its own ``patient_id`` / ``clinician_id`` (parsed at
        the

        boundary by Pydantic) and ``idempotency_key``; the URL key must match it.


        ``candidate`` is the full ``AnyWriteCandidate`` union (``WriteCandidate``
        for

        the physician-direct vital/medication kinds, ``IssueWriteCandidate`` for the

        F4b agent-proposed medical_problem/allergy kinds), so an issue write can be

        confirmed over HTTP — the two candidate shapes are distinguished by their

        disjoint ``kind`` literals. The service''s ``commit`` already accepts the

        union.'
      properties:
        candidate:
          anyOf:
          - $ref: '#/components/schemas/WriteCandidate'
          - $ref: '#/components/schemas/IssueWriteCandidate'
          title: Candidate
      required:
      - candidate
      title: ConfirmRequest
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    IssueWriteCandidate:
      description: 'A typed intake-derived issue write (medical problem / allergy)
        — F4b.


        The agent-proposed analogue of ``WriteCandidate``: same closed-surface

        discipline (exactly one payload matching ``kind``, client-generated

        ``idempotency_key``, typed principal ids), but over the issue kinds and

        defaulting to the strict ``agent_proposed_physician_confirmed`` entry mode.

        The agent may only *construct and propose* one of these; committing it

        requires the separate physician confirm transaction in

        ``writeback/service.py`` — the agent structurally cannot self-commit.'
      properties:
        allergy:
          anyOf:
          - $ref: '#/components/schemas/AllergyWrite'
          - type: 'null'
        clinician_id:
          $ref: '#/components/schemas/ClinicianId'
        entry_mode:
          $ref: '#/components/schemas/WriteEntryMode'
          default: agent_proposed_physician_confirmed
        idempotency_key:
          maxLength: 128
          minLength: 1
          title: Idempotency Key
          type: string
        kind:
          enum:
          - medical_problem
          - allergy
          title: Kind
          type: string
        medical_problem:
          anyOf:
          - $ref: '#/components/schemas/MedicalProblemWrite'
          - type: 'null'
        patient_id:
          $ref: '#/components/schemas/PatientId'
      required:
      - kind
      - patient_id
      - clinician_id
      - idempotency_key
      title: IssueWriteCandidate
      type: object
...
    MedicalProblemWrite:
      description: 'A medical-problem (issue) list entry, appended as a new list row.


        ``title`` is the intake-derived problem string the physician confirms —

        never free prose beyond it. ``begdate`` is a ``YYYY-MM-DD`` string (the

        Standard API''s format); its format is checked deterministically by the

        write verification step, not coerced here.'
      properties:
        begdate:
          description: Onset/entry date, YYYY-MM-DD.
          minLength: 1
          title: Begdate
          type: string
        diagnosis:
          anyOf:
          - type: string
          - type: 'null'
          description: Optional '<codetype>:<code>' association.
          title: Diagnosis
        title:
          minLength: 1
          title: Title
          type: string
      required:
      - title
      - begdate
      title: MedicalProblemWrite
      type: object
    MedicationWrite:
      description: 'A medication-list entry, appended as a new list row (latest-wins).


        ``title`` is a picked/echoed drug string, never free prose. ``begdate`` is
        a

        ``YYYY-MM-DD`` string (the Standard API''s format); its format is checked

        deterministically by the verifier, not coerced here.'
      properties:
        begdate:
          description: Start date, YYYY-MM-DD.
          minLength: 1
          title: Begdate
          type: string
        diagnosis:
          anyOf:
          - type: string
          - type: 'null'
          description: Optional '<codetype>:<code>' association.
          title: Diagnosis
        enddate:
          anyOf:
          - type: string
          - type: 'null'
          description: Optional end date, YYYY-MM-DD.
          title: Enddate
        title:
          minLength: 1
          title: Title
          type: string
      required:
      - title
      - begdate
      title: MedicationWrite
      type: object
...
    PatientId:
      description: OpenEMR patient PID — positive integer only.
      properties:
        value:
          exclusiveMinimum: 0.0
          title: Value
          type: integer
      required:
      - value
      title: PatientId
      type: object
    ProposeRequest:
      description: 'A raw physician write request, parsed into a typed candidate server-side.


        ``metric``/``unit`` are required for a vital and ignored for a medication

        (whose ``raw_value`` is the picked/echoed drug title). No value is ever

        interpreted by a model — the deterministic parser owns it.


        ``clinician_id`` is optional: in ``disabled`` mode it identifies the acting

        clinician (as today); in ``smart`` mode the session cookie is authoritative

        and this field, if present, is only validated against it (mismatch → 403).'
      properties:
        clinician_id:
          anyOf:
          - exclusiveMinimum: 0.0
            type: integer
          - type: 'null'
          title: Clinician Id
        kind:
          $ref: '#/components/schemas/WriteKind'
        metric:
          anyOf:
          - type: string
          - type: 'null'
          title: Metric
        patient_id:
          exclusiveMinimum: 0.0
          title: Patient Id
          type: integer
        raw_value:
          minLength: 1
          title: Raw Value
          type: string
        unit:
          anyOf:
          - type: string
          - type: 'null'
          title: Unit
      required:
      - patient_id
      - kind
      - raw_value
      title: ProposeRequest
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
    VitalWrite:
      description: 'A single-metric vital reading, appended as a new vitals form.


        ``value`` is already parsed to a number at the system boundary (parse, don''t

        validate): constructing a ``VitalWrite`` with a non-numeric value or an

        unknown metric fails here, which is the first deterministic gate. Semantic

        checks (unit-matches-metric, physiologic plausibility) are the verifier''s
        job.'
      properties:
        metric:
          $ref: '#/components/schemas/WritableMetric'
        unit:
          description: Physician-supplied unit; verified against the metric.
          minLength: 1
          title: Unit
          type: string
        value:
          title: Value
          type: number
      required:
      - metric
      - value
      - unit
      title: VitalWrite
      type: object
    WritableMetric:
      description: 'The closed set of vital-sign metrics a physician may write.


        Deliberately small: only OpenEMR''s Standard-API *vitals* columns are

        directly writable as Observation-like data (labs have no create endpoint —

        see the feasibility brief). Extend only with a matching column mapping in

        ``fhir/write_client.py`` and a plausibility spec in ``verification/writes.py``

        — the exhaustive ``match`` in each will fail to compile until you do.'
      enum:
      - heart_rate
      - spo2
      - systolic_bp
      - diastolic_bp
      - respiratory_rate
      - temperature
      - weight
      - height
      title: WritableMetric
      type: string
    WriteCandidate:
      description: 'A parsed, typed write request over the closed writable surface.


        Carries exactly one payload (``vital`` xor ``medication``) matching ``kind``.

        ``kind`` is deliberately the *direct* subset — the vital/medication verifier

        in ``verification/writes.py`` matches exhaustively over it; the issue kinds

        live on ``IssueWriteCandidate``. ``idempotency_key`` is client-generated so

        a retried/double-clicked confirm cannot create a duplicate record.

        ``patient_id`` / ``clinician_id`` scope the write; the route enforces

        ``is_authorized`` before a candidate is trusted.'
      properties:
        clinician_id:
          $ref: '#/components/schemas/ClinicianId'
        entry_mode:
          $ref: '#/components/schemas/WriteEntryMode'
          default: human_direct
        idempotency_key:
          maxLength: 128
          minLength: 1
          title: Idempotency Key
          type: string
        kind:
          enum:
          - vital
          - medication
          title: Kind
          type: string
        medication:
          anyOf:
          - $ref: '#/components/schemas/MedicationWrite'
          - type: 'null'
        patient_id:
          $ref: '#/components/schemas/PatientId'
        vital:
          anyOf:
          - $ref: '#/components/schemas/VitalWrite'
          - type: 'null'
      required:
      - kind
      - patient_id
      - clinician_id
      - idempotency_key
      title: WriteCandidate
      type: object
    WriteEntryMode:
      description: 'How the value reached the record — the physician-attribution surface.


        ``human_direct`` is Phase 1: the physician typed the value themselves, so
        an

        out-of-range value is a soft, overridable warning.

        ``agent_proposed_physician_confirmed`` is the live agent path (F4b): the

        agent drafts a typed candidate, a *separate* physician confirm transaction

        commits it, and the audit row attributes the write to this mode. The write

        verifier treats it as the strict mode that hard-blocks out-of-range values.'
      enum:
      - human_direct
      - agent_proposed_physician_confirmed
      title: WriteEntryMode
      type: string
    WriteKind:
      description: 'What kind of record a candidate creates.


        ``vital`` / ``medication`` are the physician-direct kinds (Phase 1,

        ``WriteCandidate``); ``medical_problem`` / ``allergy`` are the

        intake-derived, agent-proposed kinds (F4b, ``IssueWriteCandidate``) that

        flow through the same propose→confirm gate but require an explicit

        physician confirm to commit.'
      enum:
      - vital
      - medication
      - medical_problem
      - allergy
      title: WriteKind
      type: string
info:
  title: Clinical Co-Pilot
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /v1/writes:
    post:
      operationId: propose_write_v1_writes_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProposeRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Propose Write V1 Writes Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Propose a physician write and get the echo-back to confirm
      tags:
      - writes
  /v1/writes/{idempotency_key}/confirm:
    post:
      operationId: confirm_write_v1_writes__idempotency_key__confirm_post
      parameters:
      - in: path
        name: idempotency_key
        required: true
        schema:
          maxLength: 128
          minLength: 1
          title: Idempotency Key
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Confirm Write V1 Writes  Idempotency Key  Confirm
                  Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Confirm and commit a previously proposed write (append-only)
      tags:
      - writes